requirements.txt
fastapi==0.110.2
python-dotenv==1.0.1
uvicorn==0.29.0
openai==1.58.1
PyMuPDF==1.25.1

nltk~=3.9.1
llama-index-core~=0.12.8
psycopg2-binary~=2.9.10